Improvement: Rebase certificate logic into proper class #1295
